+++ CHANGELOG.md
@@ -23,10 +23,12 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #639 handle expired or incorrect policies when sending notifications
 - #786 Added authorization as admin for submodel api & registry api
 - #884 Upgraded tractionBatteryCode from 1.0.0 to 2.0.0
+- #1009 reimplemented retry request logic for notification approval
 - #786 Added alternative port (only accessible within same cluster) for application which is used for unsecured API endpoints.
 - #786 Introduced internal url for notification contracts.
 - #994 improved bpn edc configuration view uux
 - #1082 fix update of parts when synchronizing with IRS
+
 
 ### Added
 - #832 added policymanagement list view, creator and editor